Fix target language in translation file template
The translation file template was configured for a target language. This
prevented Linguist from showing the selection dialog when the file was
opened after copying it to a language-specific version. Not explicitly
querying the language then could have led translators to not setting a
target language and just keep the default which would have been wrong.